postgis: set "id" as BIGSERIAL
When the "id" attribute is autocreated as SERIAL it sets a relative
easy-to-reach maxint value of 2147483647.

Set the default auto-creation to BIGSERIAL (4->8 bytes), to move
the maxint target to a way higher and more difficult to reach
number.

At Wikimedia we encountered this issue when dealing with a sudden
failure of imposm, more details from (onward):

https://phabricator.wikimedia.org/T389462#10657466